- Moved the map_candidates_in_ballots_to_choices from the ballot.py file to the serializers.py file and the BallotSerializer class as a staticmethod 	- Updated calls to it from test_ballot.py
